add layout 3 report
