+++ client/pages/baocao/sanxuat.vue
@@ -1,0 +1,16 @@
+<template>
+    <div class="container">
+      <SanXuatComponent />
+    </div>
+  
+  </template>
+  
+  <script>
+import SanXuatComponent from '../../components/baocao/SanXuatComponent.vue';
+
+  export default {
+    name: "homepage",
+    components: { SanXuatComponent }
+}
+  </script>
+
