--- client/pages/baocaosanxuat.vue
@@ -1,15 +1,0 @@
-<template>
-    <div class="container">
-      <BaoCaoSanXuatComponent/>
-    </div>
-  
-  </template>
-  
-  <script>
-import BaoCaoSanXuatComponent from '../components/BaoCaoSanXuatComponent.vue';
-  export default {
-    name: "homepage",
-    components: { BaoCaoSanXuatComponent }
-}
-  </script>
-
