+++ client/pages/index.vue
@@ -6,7 +6,8 @@
 </template>
 
 <script>
-import HomePageComponent from "../components/HomePageComponent.vue";
+import HomePageComponent from "../components/baocao/HomePageComponent.vue";
+
 export default {
     name: "homepage",
     components: { HomePageComponent }
